requirements.txt
paddleocr==3.1.0
opencv-python==4.12.0.88
openai==1.98.0
scikit-learn==1.7.1
pandas==2.3.1
numpy==2.2.6
scipy==1.16.1
PyAutoGUI==0.9.54
pillow==11.3.0
pydantic==2.11.7
